Fix incorrect time in JWT

+++ src/main/java/nl/inholland/bank/models/Token.java
@@ -1,4 +1,4 @@
 package nl.inholland.bank.models;
 
-public record Token(String jwt, int expiresAt) {
+public record Token(String jwt, long expiresAt) {
 }
